requirements.txt
biopython>=1.84
matplotlib>=3.8
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
seaborn>=0.13
torch>=2.2
tqdm>=4.66
